Add dummy calls to base case for escape_json_naive
Without it, the compiler is too smart for itself and breaks the optimization.

This makes sure the clauses above (the one between escapes) and the base one
are different. If compiler sees the same clauses, it will optimize the
select_value instruction and introduce gaps. Because of those gaps, though,
the loader is not able to optimize the select_value into a jump table.