Adds float to decimal decoding option
Floats in elixir/erlang are notoriously known for being prone to
precision errors, which can be dangerous when dealing with currencies,
money, etc.

All issues that Decimal aims to mitigate.

This commit adds the option for floats to be immediately converted to
Decimals without going through float conversion.

The approach was heavily inspired on the decoding options for
strings and keys.