Add an option to decode JSON objects preserving order
This adds a new Jason.OrderedObject struct that can be used to store
a keyword (with non-atom keys) and properly implement protocols on it.

It's further used when the option `objects: :ordered_objects` is provided
for the decoder to replace native maps.

No performance impact when the option is not used was measured.

Co-authored-by: Alex Lopez <alex.lopez-zorzano@betterdoc.de>